app/tasks/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-dotenv==1.1.0
celery==5.5.1
redis==5.2.1
SQLAlchemy==2.0.40
asyncpg==0.30.0
psycopg2-binary==2.9.3
